+++ src/app/components/file-browser/terra-base-private-storage.interface.ts
@@ -1,0 +1,6 @@
+import { TerraBaseStorageService } from './terra-base-storage.interface';
+
+export abstract class TerraBasePrivateStorageService extends TerraBaseStorageService
+{
+    public abstract downloadFile(key:string):void;
+}