+++ src/app/components/forms/dynamic-form/enum/terra-control-type.enum.ts
@@ -7,10 +7,14 @@ export enum TerraControlTypeEnum
     CONDITIONAL_CONTAINER = 'conditionalContainer',
     DATE_PICKER = 'datePicker',
     HORIZONTAL_CONTAINER = 'horizontalContainer',
+    INPUT_FILE = 'inputFile',
     INPUT_TEXT = 'inputText',
     INPUT_TEXT_AREA = 'inputTextArea',
     INPUT_NUMBER = 'inputNumber',
     INPUT_DOUBLE = 'inputDouble',
     SELECT_BOX = 'selectBox',
-    VERTICAL_CONTAINER = 'verticalContainer'
+    VERTICAL_CONTAINER = 'verticalContainer',
+    CATEGORY_PICKER = 'categoryPicker',
+    COLOR_PICKER = 'colorPicker',
+    MULTI_CHECK_BOX = 'multiCheckBox'
 }