+++ src/app/components/forms/suggestion-box/data/terra-suggestion-box.interface.ts
@@ -1,8 +1,14 @@
+import { TerraBaseData } from '../../../data/terra-base.data';
+
 export interface TerraSuggestionBoxValueInterface
 {
-    value:any;
-    //i18n from locale_de and locale_en
-    caption:any;
+    /** The value of the suggestion value */
+    value:number | string | TerraBaseData;
+    // i18n from locale_de and locale_en
+    caption:string;
+    /** Optional icon for suggestion value */
     icon?:string;
+    /** Optional image for suggestion value (only the src-attribute) */
+    imgsrc?:string;
     active?:boolean;
 }