+++ src/app/components/layouts/tag/data/terra-tag.interface.ts
@@ -4,15 +4,22 @@
  */
 export interface TerraTagInterface
 {
-    isTagged?:boolean
+    isTagged?:boolean;
     isTaggable?:boolean;
+
     /**
      * @deprecated will be removed in next major release. Please use badge instead.
      */
     caption?:string;
+
     /**
      *  will be mandatory in next major release.
      */
     badge?:string;
     customClass?:string;
+
+    /**
+     * The color of the tag
+     */
+    color?:string;
 }