+++ src/app/components/navigator/example/terra-navigator.component.example.ts
@@ -1,6 +1,6 @@
 import {
     Component
-} from "@angular/core";
+} from '@angular/core';
 
 @Component({
     selector: 'terra-navigator-example',