+++ src/app/terra-components.module.ts
@@ -81,8 +81,11 @@ import { TerraBaseService } from './service/terra-base.service';
 import { TerraUrlParamsDecorator } from './service/data/terra-url-params-decorator.service';
 import { TerraAlertComponent } from './components/alert/terra-alert.component';
 import { TerraConverterHelper } from './helpers/terra-converter.helper';
+import { TerraCategoryPickerBaseService } from './components/category-picker/service/terra-category-picker-base.service';
+import { CategoryTreeConfig } from './components/category-picker/config/category-tree.config';
 import { TerraJsonToFormFieldService } from './components/forms/dynamic-form/service/terra-json-to-form-field.service';
-//### import of terra-component examples
+
+// ### import of terra-component examples
 import { TerraButtonComponentExample } from './components/buttons/button/example/terra-button.component.example';
 import { TerraAlertComponentExample } from './components/alert/example/terra-alert.component.example';
 import { TerraCardComponentExample } from './components/layouts/card/example/terra-card.component.example';
@@ -117,7 +120,9 @@ import { TerraNoResultNoticeComponentExample } from './components/no-result/exam
 import { TerraButtonWithOptionsComponentExample } from './components/buttons/button-with-options/example/terra-button-with-options.component.example';
 import { TerraInputComponentExample } from './components/forms/input/example/terra-input.component.example';
 import { TerraFilterComponentExample } from './components/filter/example/terra-filter.component.example';
-import { TerraDataTableContextMenuComponentExample } from './components/tables/data-table/context-menu/example/terra-data-table-context-menu.component.example';
+import {
+    TerraDataTableContextMenuComponentExample
+} from './components/tables/data-table/context-menu/example/terra-data-table-context-menu.component.example';
 import { TerraNoteEditorComponentExample } from './components/editors/note-editor/example/terra-note-editor.component.example';
 import { TerraNoteComponentExample } from './components/note/example/terra-note.component.example';
 import { TerraNodeTreeComponentExample } from './components/tree/node-tree/example/terra-node-tree.component.example';
@@ -125,6 +130,9 @@ import { TerraDynamicFormComponent } from './components/forms/dynamic-form/terra
 import { TerraDynamicSwitchComponent } from './components/forms/dynamic-form/dynamic-switch/terra-dynamic-switch.component';
 import { TerraDynamicFormService } from './components/forms/dynamic-form/service/terra-dynamic-form.service';
 import { TerraFormFieldControlService } from './components/forms/dynamic-form/service/terra-form-field-control.service';
+import { TerraMultiCheckBoxComponent } from './components/forms/multi-check-box/terra-multi-check-box.component';
+import { TerraMultiCheckBoxComponentExample } from './components/forms/multi-check-box/example/terra-multi-check-box.component.example';
+import { TerraCategoryPickerComponent } from './components/category-picker/terra-category-picker.component';
 
 @NgModule({
     declarations:    [
@@ -157,6 +165,7 @@ import { TerraFormFieldControlService } from './components/forms/dynamic-form/se
         TerraSplitViewComponent,
         TerraFilterComponent,
         TerraMultiSelectBoxComponent,
+        TerraMultiCheckBoxComponent,
         TerraDynamicComponentLoaderComponent,
         TerraDynamicModuleLoaderComponent,
         TerraTileBoxComponent,
@@ -179,10 +188,12 @@ import { TerraFormFieldControlService } from './components/forms/dynamic-form/se
         TerraNoteComponent,
         TerraNodeTreeComponent,
         TerraNodeComponent,
+        TerraCategoryPickerComponent,
         TerraDynamicFormComponent,
         TerraDynamicSwitchComponent,
+        TerraCategoryPickerComponent,
 
-        //### declarations of terra-components examples
+        // ### declarations of terra-components examples
         TerraButtonComponentExample,
         TerraAlertComponentExample,
         TerraCardComponentExample,
@@ -198,6 +209,7 @@ import { TerraFormFieldControlService } from './components/forms/dynamic-form/se
         TerraTextInputComponentExample,
         TerraCheckboxComponentExample,
         TerraMultiSelectBoxComponentExample,
+        TerraMultiCheckBoxComponentExample,
         TerraRadioButtonComponentExample,
         TerraSelectBoxComponentExample,
         TerraSliderComponentExample,
@@ -250,6 +262,7 @@ import { TerraFormFieldControlService } from './components/forms/dynamic-form/se
         TerraSplitViewComponent,
         TerraFilterComponent,
         TerraMultiSelectBoxComponent,
+        TerraMultiCheckBoxComponent,
         TerraTileBoxComponent,
         TerraTileBoxPanelComponent,
         TerraSuggestionBoxComponent,
@@ -267,7 +280,8 @@ import { TerraFormFieldControlService } from './components/forms/dynamic-form/se
         TerraNoteEditorComponent,
         TerraNoteComponent,
         TerraNodeTreeComponent,
-        TerraNodeComponent
+        TerraNodeComponent,
+        TerraCategoryPickerComponent
     ],
     exports:         [
         TerraAlertPanelComponent,
@@ -298,6 +312,7 @@ import { TerraFormFieldControlService } from './components/forms/dynamic-form/se
         TerraSplitViewComponent,
         TerraFilterComponent,
         TerraMultiSelectBoxComponent,
+        TerraMultiCheckBoxComponent,
         TerraDynamicComponentLoaderComponent,
         TerraDynamicModuleLoaderComponent,
         TerraTileBoxComponent,
@@ -319,11 +334,12 @@ import { TerraFormFieldControlService } from './components/forms/dynamic-form/se
         TerraNoteEditorComponent,
         TerraNoteComponent,
         TerraNodeTreeComponent,
+        TerraCategoryPickerComponent,
         TerraDynamicFormComponent,
         TerraDynamicSwitchComponent,
+        TerraCategoryPickerComponent,
 
-
-        //### declarations of terra-components examples
+        // ### declarations of terra-components examples
         TerraButtonComponentExample,
         TerraAlertComponentExample,
         TerraCardComponentExample,
@@ -339,6 +355,7 @@ import { TerraFormFieldControlService } from './components/forms/dynamic-form/se
         TerraTextInputComponentExample,
         TerraCheckboxComponentExample,
         TerraMultiSelectBoxComponentExample,
+        TerraMultiCheckBoxComponentExample,
         TerraRadioButtonComponentExample,
         TerraSelectBoxComponentExample,
         TerraSliderComponentExample,
@@ -387,7 +404,7 @@ import { TerraFormFieldControlService } from './components/forms/dynamic-form/se
         TerraNavigatorSplitViewConfig,
         TerraFrontendStorageService,
         TerraFileBrowserService,
-        TerraConverterHelper
+        TerraConverterHelper,
     ],
     bootstrap:       [
         TerraComponentsComponent
@@ -427,6 +444,7 @@ export class TerraComponentsModule
                 TerraUrlParamsDecorator,
                 TerraFrontendStorageService,
                 TerraAlertComponent,
+                CategoryTreeConfig,
                 TerraJsonToFormFieldService
             ]
         };